[ext] reduce shell icon refreshing overhead
* Shell extension keeps a repos information timestamp, applet only
returns repos info when timestamp is outdated
* Only trigger a shell icon refresh when repo status really changed

+++ src/seafile-applet.cpp
@@ -551,7 +551,7 @@ void SeafileApplet::updateReposPropertyForHttpSync()
     }
 
     const std::vector<Account>& accounts = account_mgr_->accounts();
-    for (int i = 0; i < repos.size(); i++) {
+    for (size_t i = 0; i < repos.size(); i++) {
         const LocalRepo& repo = repos[i];
         QString repo_server_url;
         QString relay_addr;
@@ -564,7 +564,7 @@ void SeafileApplet::updateReposPropertyForHttpSync()
         if (rpc_client_->getRepoProperty(repo.id, kRepoRelayAddrProperty, &relay_addr) < 0) {
             continue;
         }
-        for (int i = 0; i < accounts.size(); i++) {
+        for (size_t i = 0; i < accounts.size(); i++) {
             const Account& account = accounts[i];
             if (account.serverUrl.host() == relay_addr) {
                 QUrl url(account.serverUrl);
